api/requirements.txt
litestar[standard]
pydantic
open-groceries
pymongo
python-dotenv
minio